Compute command type during evaluation
The Parser has now a mutable member variable CommandType that is
updated as the current command is being evaluated. This refactor
relieves the parser from having to evaluate the command twice:
once to check if it is a valid command and then again when the
command type was being requested.